+++ lib/mongoid/fields/internal/foreign_keys/array.rb
@@ -22,7 +22,7 @@ module Mongoid #:nodoc:
           # @since 2.4.0
           def add_atomic_changes(document, key, mods, new, old)
             if new.any? && old.any?
-              mods[key] = new
+              mods[key] = document.attributes[key]
             elsif new.any?
               document.atomic_array_add_to_sets[key] = new
             elsif old.any?