+++ spec/unit/mongoid/fields/internal/array_spec.rb
@@ -6,6 +6,34 @@ describe Mongoid::Fields::Internal::Array do
     described_class.instantiate(:test, :type => Array)
   end
 
+  describe "#add_atomic_changes" do
+
+    let(:person) do
+      Person.new
+    end
+
+    let(:mods) do
+      {}
+    end
+
+    before do
+      person.aliases = [ "007", "008" ]
+    end
+
+    context "when adding and removing" do
+
+      before do
+        field.add_atomic_changes(
+          person, "aliases", mods, [ "008" ], [ "009" ]
+        )
+      end
+
+      it "adds the current to the modifications" do
+        mods["aliases"].should eq([ "007", "008" ])
+      end
+    end
+  end
+
   describe "#cast_on_read?" do
 
     it "returns false" do